+++ llama_index/llms/huggingface.py
@@ -387,7 +387,7 @@ class HuggingFaceInferenceAPI(LLM):
             "The model to run inference with. Can be a model id hosted on the Hugging"
             " Face Hub, e.g. bigcode/starcoder or a URL to a deployed Inference"
             " Endpoint. Defaults to None, in which case a recommended model is"
-            " automatically selected for the task."
+            " automatically selected for the task (see Field below)."
         ),
     )
     token: Union[str, bool, None] = Field(
@@ -415,6 +415,13 @@ class HuggingFaceInferenceAPI(LLM):
     )
     cookies: Dict[str, str] = Field(
         default=None, description="Additional cookies to send to the server."
+    )
+    task: Optional[str] = Field(
+        default=None,
+        description=(
+            "Optional task to pick Hugging Face's recommended model, used when"
+            " model_name is left as default of None."
+        ),
     )
     _sync_client: "InferenceClient" = PrivateAttr()
     _async_client: "AsyncInferenceClient" = PrivateAttr()
@@ -464,7 +471,6 @@ class HuggingFaceInferenceAPI(LLM):
         Args:
             kwargs: See the class-level Fields.
         """
-        super().__init__(**kwargs)  # Populate pydantic Fields
         try:
             from huggingface_hub import (
                 AsyncInferenceClient,
@@ -474,8 +480,18 @@ class HuggingFaceInferenceAPI(LLM):
         except ModuleNotFoundError as exc:
             raise ImportError(
                 f"{type(self).__name__} requires huggingface_hub with its inference"
-                " extras, please run `pip install huggingface_hub[inference]`."
+                " extra, please run `pip install huggingface_hub[inference]>=0.19.0`."
             ) from exc
+        if kwargs.get("model_name") is None:
+            task = kwargs.get("task", "")
+            # NOTE: task being None or empty string leads to ValueError,
+            # which ensures model is present
+            kwargs["model_name"] = InferenceClient.get_recommended_model(task=task)
+            logger.debug(
+                f"Using Hugging Face's recommended model {kwargs['model_name']}"
+                f" given task {task}."
+            )
+        super().__init__(**kwargs)  # Populate pydantic Fields
         self._sync_client = InferenceClient(**self._get_inference_client_kwargs())
         self._async_client = AsyncInferenceClient(**self._get_inference_client_kwargs())
         self._get_model_info = model_info