+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - `OpenAIAgent` parallel function calling (#8738)
 
+### New Features
+
+- Properly supporting Hugging Face recommended model (#8784)
+
 ### Bug Fixes / Nits
 
 - Fixed missing import for `embeddings.__all__` (#8779)